% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.fr_col}
\alias{is.fr_col}
\title{Test if an Object is an fr_col}
\usage{
is.fr_col(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"fr_col"}, \code{FALSE} otherwise.
}
\description{
Test if an Object is an fr_col
}
\examples{
col <- fr_col(label = "Treatment", align = "center")
is.fr_col(col)       # TRUE
is.fr_col("text")    # FALSE

}
