% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{is.fr_spec}
\alias{is.fr_spec}
\title{Test if an Object is an fr_spec}
\usage{
is.fr_spec(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"fr_spec"}, \code{FALSE} otherwise.
}
\description{
Test if an Object is an fr_spec
}
\examples{
spec <- tbl_demog |> fr_table()
is.fr_spec(spec)     # TRUE
is.fr_spec(mtcars)   # FALSE

}
